♻️ runtime 프리미티브 yule_runtime 이동 + compat shim + 경로 wiring
변경 이유:
circuit_breaker / services / subprocess_supervisor 는 stdlib + 자기
sibling 외에는 의존하지 않는 저결합 프리미티브다. 의존성 규칙 (agent /
discord / memory 내부 import 금지) 을 통과하는 이 3 개만 패키지로 옮기고,
기존 import 경로는 동일 객체로 유지한다. heartbeats 는 ..agents.job_queue
를 import 하므로 제외, status.py 는 범위 외로 남긴다.

주요 변경 사항:
- circuit_breaker.py / services.py / subprocess_supervisor.py 를
  packages/runtime/src/yule_runtime/ 로 이동 (subprocess_supervisor 의
  .circuit_breaker / .services 상대 import 는 sibling 으로 유지)
- 기존 경로 src/yule_orchestrator/runtime/<name>.py 를 thin compat shim
  으로 교체 — sys.modules alias 방식으로 옛 경로를 새 모듈에 매핑해
  객체 동일성 보장 (services.PROFILES monkeypatch / importlib.reload 도
  동일 객체에서 동작)
- 루트 pyproject [tool.setuptools.packages.find] where 에
  "packages/runtime/src" 추가

비고:
- CLI 표면 (yule runtime up/status/down) 및 status_cli / circuit_cli /
  status_poster / status_summary 의 import 는 shim 으로 그대로 resolve.
- 의존성 규칙 위반/coupling 으로 남긴 모듈은 packages/runtime/README.md
  TODO 참조.
