+++ pyproject.toml
@@ -18,4 +18,4 @@ dependencies = [
 yule = "yule_orchestrator.cli:main"
 
 [tool.setuptools.packages.find]
-where = ["src"]
+where = ["src", "packages/runtime/src"]
